Renamed files to match name change of base classes.
